+++ src/controllers/userController.ts
@@ -1,6 +1,12 @@
 import bcrypt from "bcrypt";
 import { Request, Response } from "express";
 import { prisma } from "../prisma";
+import { IUserInput } from "../models/user-model";
+
+// Utility function to check if the error is an instance of Error
+function isError(error: unknown): error is Error {
+  return error instanceof Error;
+}
 
 // Get all users
 export const getAllUsers = async (
@@ -11,8 +17,12 @@ export const getAllUsers = async (
     const allUsers = await prisma.user.findMany();
     res.status(200).json({ message: "All users list:", allUsers });
   } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: "Database error" });
+    if (isError(error)) {
+      console.error(error.message);
+      res.status(500).json({ error: error.message });
+    } else {
+      res.status(500).json({ error: "An unknown error occurred." });
+    }
   }
 };
 
@@ -21,27 +31,40 @@ export const createUser = async (
   req: Request,
   res: Response
 ): Promise<void> => {
-  const { name, email, password } = req.body;
+  const { name, email, password }: IUserInput = req.body;
+
   try {
+    const hashedPassword = await bcrypt.hash(password, 10);
     const result = await prisma.user.create({
       data: {
         name,
         email,
-        password: await bcrypt.hash(password, 10),
+        password: hashedPassword,
       },
     });
     res.status(201).json({ message: "User created", result });
   } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: "Database error" });
+    if (isError(error)) {
+      console.error(error.message);
+      res.status(500).json({ error: error.message });
+    } else {
+      res.status(500).json({ error: "An unknown error occurred." });
+    }
   }
 };
 
+// Get a user by ID
 export const getUserById = async (
   req: Request,
   res: Response
 ): Promise<void> => {
   const { id } = req.params;
+
+  if (isNaN(parseInt(id))) {
+    res.status(400).json({ error: "Invalid user ID" });
+    return;
+  }
+
   try {
     const user = await prisma.user.findUnique({
       where: {
@@ -49,23 +72,34 @@ export const getUserById = async (
       },
     });
 
-    if (user === null) {
+    if (!user) {
       res.status(404).json({ error: "User not found" });
       return;
     }
+
     res.status(200).json({ message: "Get user:", user });
   } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: "Database error" });
+    if (isError(error)) {
+      console.error(error.message);
+      res.status(500).json({ error: error.message });
+    } else {
+      res.status(500).json({ error: "An unknown error occurred." });
+    }
   }
 };
 
+// Delete a user by ID
 export const deleteUser = async (
   req: Request,
   res: Response
 ): Promise<void> => {
   const { id } = req.params;
-  console.log(id);
+
+  if (isNaN(parseInt(id))) {
+    res.status(400).json({ error: "Invalid user ID" });
+    return;
+  }
+
   try {
     const result = await prisma.user.delete({
       where: {
@@ -73,14 +107,18 @@ export const deleteUser = async (
       },
     });
 
-    console.log(result);
-    if (result === null) {
+    if (!result) {
       res.status(404).json({ error: "User not found" });
       return;
     }
+
     res.status(204).send(`User with ID ${id} successfully deleted`);
   } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: "Database error" });
+    if (isError(error)) {
+      console.error(error.message);
+      res.status(500).json({ error: error.message });
+    } else {
+      res.status(500).json({ error: "An unknown error occurred." });
+    }
   }
 };
